fix(ci): python-full-coverage — extract setup-commands test harness under the 1500-line limit
tests/unit/cli/test_setup_commands.py had grown to 1553 lines, tripping
test_first_party_code_files_stay_under_line_limit. Move the shared
scaffolding (runner handle, hermetic `harness` fixture, check-result
builders, source-checkout helper) into a new
tests/unit/cli/_setup_commands_harness.py and import it back, mirroring
the existing test_init_parts/_bootstrap_helper pattern. The test module
drops to 1436 lines; no test behavior changes.
